Rework modules, add types, split out command actions, needs parser help
