+++ cabal.project
@@ -1,0 +1,2 @@
+packages: .
+
